requirements.txt
pycryptodome==3.22.0
pytest==8.3.2
flake8==7.1.1
